\name{mantaExists}
\alias{mantaExists}
\title{Tests to see if a Manta file or subdirectory exists}
\usage{
  mantaExists(mantapath)
}
\arguments{
  \item{mantapath}{character, required.}
}
\value{
  logical.
}
\description{
  Tests to see if a Manta file or subdirectory exists
}
\keyword{Manta,}
\keyword{manta}

