\name{mantaJob.input}
\alias{mantaJob.input}
\title{mantaJob.input returns list of input objects given Manta job identifier}
\usage{
  mantaJob.input(jobid)
}
\arguments{
  \item{jobid}{character required. Manta job identifier
  such as "70c30bab-873b-66da-ebc8-ced12bd35ac4" or use
  mantaJob.last() to fetch the jobid of the last manta Job
  run on the service e.g. mantaJob.errors(mantaJob.last())}
}
\description{
  mantaJob.input returns list of input objects given Manta
  job identifier
}
\keyword{Manta,}
\keyword{manta}

