\name{mantaJobs.running}
\alias{mantaJobs.running}
\title{Lists identifiers of any running Manta jobs.}
\usage{
mantaJobs.running()
}
\description{
Lists identifiers of any running Manta jobs.
}
\examples{
\dontrun{
## List any running Manta jobs:
mantaJobs.running()
}
}
\seealso{
Other mantaJobs: \code{\link{mantaJob.cancel}};
\code{\link{mantaJob.done}};
\code{\link{mantaJob.errors.stderr}};
\code{\link{mantaJob.errors}};
\code{\link{mantaJob.failures}};
\code{\link{mantaJob.inputs}};
\code{\link{mantaJob.launch}};
\code{\link{mantaJob.outputs.cat}};
\code{\link{mantaJob.outputs}};
\code{\link{mantaJob.setup}};
\code{\link{mantaJob.status}};
\code{\link{mantaJobs.tail}}; \code{\link{mantaJobs}};
\code{\link{mantaMap}}; \code{\link{mantaReduce}}
}
\keyword{Manta}

