\name{mantaJobs.tail}
\alias{mantaJobs.tail}
\title{Returns identifier of last run Manta job identifier, or from offset n up from end of list.}
\usage{
  mantaJobs.tail(n = 1)
}
\arguments{
  \item{n}{integer. Index into most recently run Jobs list.
  n = 1 default is last Manta Job run. 2 is second last,
  and so on.}
}
\description{
  Returns identifier of last run Manta job identifier, or
  from offset n up from end of list.
}
\keyword{Manta,}
\keyword{manta}

