\name{mantaLoad.ws}
\alias{mantaLoad.ws}
\title{Loads current R workspace from Manta R workspace directory}
\usage{
  mantaLoad.ws(envir = parent.frame())
}
\arguments{
  \item{envir}{optional. Enviroment into which workspace is
  loaded.}
}
\value{
  logical.
}
\description{
  Loads current R workspace from Manta R workspace
  directory
}
\keyword{Manta,}
\keyword{manta}

