\name{mantaMetadata}
\alias{mantaMetadata}
\title{Returns the HTTP HEAD metadata for a Manta object or subdirectory}
\usage{
  mantaMetadata(mantapath)
}
\arguments{
  \item{mantapath}{string, required.}
}
\value{
  character array of HTTP HEAD contents.
}
\description{
  Returns the HTTP HEAD metadata for a Manta object or
  subdirectory
}
\keyword{Manta,}
\keyword{manta}

