\name{mantaMkdir}
\alias{mantaMkdir}
\title{Makes a Manta subdirectory, optionally with parent directories.}
\usage{
  mantaMkdir(mantapath, p = FALSE)
}
\arguments{
  \item{mantapath}{string, required.}

  \item{p}{logical, optional. Make all the parent
  directories too}
}
\description{
  Makes a Manta subdirectory, optionally with parent
  directories.
}
\keyword{Manta,}
\keyword{manta}

