\name{mantaRmdir}
\alias{mantaRmdir}
\title{Removes Manta subdirectory.}
\usage{
  mantaRmdir(mantapath)
}
\arguments{
  \item{mantapath}{string, required.}
}
\value{
  logical, TRUE if object successfully removed.
}
\description{
  Specifiy the full manta path or the just subdirectory
  name from current working Manta directory. Returns TRUE
  if object successfully removed
}
\keyword{Manta,}
\keyword{manta}

