\name{mantaSetwd.public}
\alias{mantaSetwd.public}
\title{Sets current Manta working directory to \code{~~/public}}
\usage{
mantaSetwd.public()
}
\description{
See \code{mantaSetwd}.
}
\seealso{
Other Directory: \code{\link{mantaGetwd}};
\code{\link{mantaMkdir}}; \code{\link{mantaRmdir}};
\code{\link{mantaRm}}; \code{\link{mantaSetwd.reports}};
\code{\link{mantaSetwd.stor}}; \code{\link{mantaSetwd}};
\code{\link{mantaSnapln}}
}
\keyword{Manta}

