\name{mantaSetwd.reports}
\alias{mantaSetwd.reports}
\title{Sets current Manta working directory to \code{~~/reports}}
\usage{
mantaSetwd.reports()
}
\description{
See \code{mantaSetwd}.
}
\seealso{
Other Directory: \code{\link{mantaGetwd}};
\code{\link{mantaMkdir}}; \code{\link{mantaRmdir}};
\code{\link{mantaRm}}; \code{\link{mantaSetwd.public}};
\code{\link{mantaSetwd.stor}}; \code{\link{mantaSetwd}};
\code{\link{mantaSnapln}}
}
\keyword{Manta}

