\name{mantaSetwd.stor}
\alias{mantaSetwd.stor}
\title{Sets current Manta working directory to \code{~~/stor}}
\usage{
  mantaSetwd.stor()
}
\description{
  See \code{mantaSetwd}.
}
\seealso{
  Other Directory: \code{\link{mantaGetwd}},
  \code{\link{mantaMkdir}}, \code{\link{mantaRm}},
  \code{\link{mantaRmdir}}, \code{\link{mantaSetwd}},
  \code{\link{mantaSetwd.public}},
  \code{\link{mantaSetwd.reports}},
  \code{\link{mantaSnapln}}
}
\keyword{Manta,}
\keyword{manta}

