\name{mantaSetwd.ws}
\alias{mantaSetwd.ws}
\title{Sets the current Manta working directory to the R workspace e.g.  ~~/stor/R-3.0.1/}
\usage{
  mantaSetwd.ws()
}
\value{
  logical.
}
\description{
  Sets the current Manta working directory to the R
  workspace e.g.  ~~/stor/R-3.0.1/
}
\keyword{Manta,}
\keyword{manta}

