\name{mantaunixstyle}
\alias{mantaunixstyle}
\title{mantaLs and mantaFind callback - Unix like listings
mimic of ln -o but some of this is static}
\usage{
  mantaunixstyle(line)
}
\arguments{
  \item{line}{R structured directory line}
}
\description{
  mantaLs and mantaFind callback - Unix like listings mimic
  of ln -o but some of this is static
}

