\name{mantaCat}
\alias{mantaCat}
\title{Retrieves object from Manta and uses cat() to print contents to the R console.}
\usage{
  mantaCat(mantapath)
}
\arguments{
  \item{mantapath}{string, required.}
}
\description{
  Retrieves object from Manta and uses cat() to print
  contents to the R console.
}
\keyword{Manta,}
\keyword{at}
\keyword{binary}
\keyword{data.}
\keyword{files,}
\keyword{for}
\keyword{intended}
\keyword{is}
\keyword{manta}
\keyword{mantaCat}
\keyword{on}
\keyword{own}
\keyword{risk}
\keyword{text}
\keyword{use}
\keyword{your}

