\name{mantaExists}
\alias{mantaExists}
\title{Tests to see if a Manta object or subdirectory exists.}
\usage{
  mantaExists(mantapath, d = FALSE)
}
\arguments{
  \item{mantapath}{character, required.}

  \item{d}{logical. Set TRUE to confirm that entity
  specified in mantapath exists and is a directory.}
}
\value{
  logical.
}
\description{
  Tests to see if a Manta object or subdirectory exists.
}
\keyword{Manta,}
\keyword{manta}

