\name{mantaJobs.tail}
\alias{mantaJobs.tail}
\title{Returns the last run Manta job identifiers, sorted by time.}
\usage{
  mantaJobs.tail(n = 1)
}
\arguments{
  \item{n}{integer. Index into most recently run Jobs list.
  n = 1 default is last Manta Job run. 2 is second last,
  and so on.}
}
\description{
  Returns the last run Manta job identifiers, sorted by
  time.
}
\keyword{Manta,}
\keyword{manta}

