\name{mantaLoad.ws}
\alias{mantaLoad.ws}
\title{Loads current R workspace from Manta ~~/stor/R-3.0.1/ workspace directory}
\usage{
  mantaLoad.ws(envir = parent.frame())
}
\arguments{
  \item{envir}{optional. Enviroment into which workspace is
  loaded.}
}
\value{
  logical.
}
\description{
  Loads current R workspace from Manta ~~/stor/R-3.0.1/
  workspace directory
}
\keyword{Manta,}
\keyword{manta}

