\name{mantaRm}
\alias{mantaRm}
\title{Removes Manta object specified by full manta path or from current
working manta directory
Returns TURE if object successfully removed}
\usage{
  mantaRm(mantapath)
}
\arguments{
  \item{mantapath}{string, required.}
}
\description{
  Removes Manta object specified by full manta path or from
  current working manta directory Returns TURE if object
  successfully removed
}
\keyword{Manta,}
\keyword{manta}

