\name{mantaRmdir}
\alias{mantaRmdir}
\title{Removes Manta subdirectory specified by full manta path or from current
working manta directory
Returns TURE if object successfully removed}
\usage{
  mantaRmdir(mantapath)
}
\arguments{
  \item{mantapath}{string, required.}
}
\description{
  Removes Manta subdirectory specified by full manta path
  or from current working manta directory Returns TURE if
  object successfully removed
}
\keyword{Manta,}
\keyword{manta}

