\name{mantaSetwd.public}
\alias{mantaSetwd.public}
\title{Sets current Manta working directory to \code{~~/public}}
\usage{
  mantaSetwd.public()
}
\description{
  See \code{mantaSetwd}.
}
\seealso{
  Other Directory: \code{\link{mantaGetwd}},
  \code{\link{mantaMkdir}}, \code{\link{mantaRm}},
  \code{\link{mantaRmdir}}, \code{\link{mantaSetwd}},
  \code{\link{mantaSetwd.reports}},
  \code{\link{mantaSetwd.stor}}, \code{\link{mantaSnapln}}
}
\keyword{Manta,}
\keyword{manta}

