\name{mantaSetwd.stor}
\alias{mantaSetwd.stor}
\title{Sets current Manta working directory to \code{~~/stor}}
\usage{
mantaSetwd.stor()
}
\description{
See \code{mantaSetwd}.
}
\seealso{
Other Directory: \code{\link{mantaGetwd}};
\code{\link{mantaMkdir}}; \code{\link{mantaRmdir}};
\code{\link{mantaRm}}; \code{\link{mantaSetwd.public}};
\code{\link{mantaSetwd.reports}}; \code{\link{mantaSetwd}};
\code{\link{mantaSnapln}}
}
\keyword{Manta}

