\name{mantaSetwd.ws}
\alias{mantaSetwd.ws}
\title{Sets the current Manta working directory to the R workspace.\cr
E.g.  \code{~~/stor/R-3.0.1/myworkstation}}
\usage{
  mantaSetwd.ws()
}
\value{
  logical.
}
\description{
  Sets the current Manta working directory to the R
  workspace.\cr E.g.  \code{~~/stor/R-3.0.1/myworkstation}
}
\keyword{Manta,}
\keyword{manta}

